+++ src/libANGLE/validationES.cpp
@@ -1455,7 +1455,7 @@ static bool ValidateDrawBase(Context *context, GLenum mode, GLsizei count, GLsiz
         if (uniformBufferSize == 0)
         {
             // Bind the whole buffer.
-            uniformBufferSize = uniformBuffer->getSize();
+            uniformBufferSize = static_cast<size_t>(uniformBuffer->getSize());
         }
 
         if (uniformBufferSize < uniformBlock->dataSize)
@@ -1629,7 +1629,7 @@ bool ValidateDrawElements(Context *context, GLenum mode, GLsizei count, GLenum t
     if (elementArrayBuffer)
     {
         uintptr_t offset = reinterpret_cast<uintptr_t>(indices);
-        if (!elementArrayBuffer->getIndexRangeCache()->findRange(type, offset, count, indexRangeOut))
+        if (!elementArrayBuffer->getIndexRangeCache()->findRange(type, static_cast<unsigned int>(offset), count, indexRangeOut))
         {
             rx::BufferImpl *bufferImpl = elementArrayBuffer->getImplementation();
             const uint8_t *dataPointer = NULL;
@@ -1642,7 +1642,7 @@ bool ValidateDrawElements(Context *context, GLenum mode, GLsizei count, GLenum t
 
             const uint8_t *offsetPointer = dataPointer + offset;
             *indexRangeOut = rx::IndexRangeCache::ComputeRange(type, offsetPointer, count);
-            elementArrayBuffer->getIndexRangeCache()->addRange(type, offset, count, *indexRangeOut);
+            elementArrayBuffer->getIndexRangeCache()->addRange(type, static_cast<unsigned int>(offset), count, *indexRangeOut);
         }
     }
     else
